fix(macro): fetch_fred 強制 value→float64,修「整數 series 觸發 SchemaError」(v19.172) (#410)
## 問題(user 截圖)
總經載入失敗,主訊息(v19.171 上浮後可見):
`SchemaError`: expected series 'value' to have type float64, got int64

## 根因
`repositories/macro_repository.py:294` 用 `pd.to_numeric(df["value"], errors="coerce")`,
對「全為整數字串」的 FRED series(如 PAYEMS / HSN1F / ICSA 等就業/住宅
啟動數),pandas dtype inference 給出 int64 而非 float64,違反
`MacroFredSchema.value = "float64"` 契約(shared/schemas.py:74)。

v19.171 修了 catastrophic propagation(per-series 失敗回空 DF),但根因仍在 —
所有「全整數」FRED series 都會踩,該指標被 silent drop。

## 修正
`.astype("float64")` 強制統一 dtype,不依賴 pandas dtype inference。

## 守衛
tests/test_schemas_phase_a.py 新增 `test_fetch_fred_integer_only_series_dtype`
mock 全整數 payload,assert dtype == "float64"。

+++ repositories/macro_repository.py
@@ -291,7 +291,12 @@ def fetch_fred(series_id: str, api_key: str, n: int = 250) -> pd.DataFrame:
         return pd.DataFrame()
     df = pd.DataFrame(obs)
     df = df[df["value"] != "."].copy()
-    df["value"] = pd.to_numeric(df["value"], errors="coerce")
+    # v19.172:強制轉 float64(不可只用 pd.to_numeric 後 dtype inference)。
+    # FRED 部分 series 全為整數(PAYEMS / HSN1F / ICSA 等就業/住宅啟動數),
+    # to_numeric 對「全整數」會回 int64,違反 MacroFredSchema "float64" 契約
+    # → pandera SchemaError → 整個 fetch_all_indicators 炸(v19.171 修了
+    # catastrophic propagation,但根因仍在此)。
+    df["value"] = pd.to_numeric(df["value"], errors="coerce").astype("float64")
     df["date"]  = pd.to_datetime(df["date"])
     # v19.60 D1：FRED API observations 已含 realtime_start 欄位（YYYY-MM-DD 字串）。
     # 缺欄或解析失敗回 NaT，呼叫端用 .get/.dropna 容錯。
